refactor: Split styles and component
